+++ frontend/src/pages/Login/LoginPage.scss
@@ -1,0 +1,124 @@
+@import '../../styles/global';
+
+.login{
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  max-width: 300px;
+  width: 100%;
+  height: auto;
+  padding: 50px 0;
+  margin: 0 auto;
+
+  h2{
+    font-size: $font-size-h1;
+    font-weight: 700;
+    margin-bottom: 1rem;
+  }
+
+  .login-container{
+    width: 100%;
+    border-bottom: 1px solid #ddd;
+
+    .login-form{
+      .input-wrap{
+        display: flex;
+        flex-direction: column;
+        margin-bottom: 5px;
+
+        .input-field{
+          &:nth-of-type(1){
+            input{
+              border-radius: 4px 4px 0 0;
+              border-bottom: 0;
+            }
+          }
+          &:nth-of-type(2){
+            input{
+              border-radius: 0 0 4px 4px ;
+            }
+          }
+        }
+      }
+      .login-btn{
+        margin-top: 1rem;
+        background-color: #333;
+      }
+    }
+    .password-reset{
+      display: flex;
+      justify-content: flex-end;
+      margin-top: 1rem;
+      font-size: $font-size-small;
+    }
+
+    .sns-login{
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      gap: 1rem;
+      width: 100%;
+      height: auto;
+      padding: 20px 0;
+      .sns-icon {
+        display: inline-block;
+        width: 50px;
+        height: 50px;
+        background-size: cover;
+        background-position: center;
+
+        &.kakao{
+          background-image: url("../../assets/images/icon-kakao.png");
+        }
+
+        &.google{
+          background-image: url("../../assets/images/icon-google.svg");
+        }
+
+        &.naver{
+          background-image: url("../../assets/images/icon-naver.png");
+        }
+
+      }
+    }
+  }
+
+  .order-container{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: 1rem;
+    width: 100%;
+    margin-top: 1rem;
+
+    .join{
+      width: 100%;
+      padding: .5rem 1rem;
+      border: 1px solid #333;
+      border-radius: 4px;
+      text-align: center;
+      font-size: $font-size-h4;
+      cursor: pointer;
+    }
+
+    .toggle-button-container{
+
+      .toggle-content{
+
+        .order-form{
+          display: flex;
+          flex-direction: column;
+          gap: .5rem;
+
+          .input-field{
+            input{
+              border-radius: 4px;
+            }
+          }
+        }
+      }
+    }
+  }
+}
