+++ frontend/src/styles/variables.scss
@@ -1,0 +1,8 @@
+/* Font Size */
+$font-size-h1: 2rem;    // 32px
+$font-size-h2: 1.5rem;  // 24px
+$font-size-h3: 1.25rem; // 20px
+$font-size-h4: 1.125rem; // 18px
+$font-size-body: 1rem;  // 16px
+$font-size-small: 0.875rem; // 14px
+$font-size-caption: 0.75rem; // 12px
